refactor(harness): generalize ACP live control adapter
